vae/requirements.txt
fenn
fenn[torch]
torchvision
matplotlib
tqdm
numpy
